requirements.txt
amqp==5.1.0
azure-core==1.23.1
azure-storage-blob==12.11.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==36.0.2
Flask==2.1.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.1
kombu==5.2.4
mysql-connector-python==8.0.26
MarkupSafe==2.1.1
msrest==0.6.21
numpy==1.21.2
oauthlib==3.2.0
pandas==1.3.3
prompt-toolkit==3.0.28
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.1
redis==3.5.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==1.4.34
typing-extensions==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.1.0
zipp==3.7.0
